refactor(cli): make completion the canonical noun group
Doctrine 1b (04_exceptions.md, amended 2026-07-07): a bare completion
command is banned; the completion GROUP with verb subcommands is the
canon. Changes:

- move the group out of main.py into scitex/cli/completion.py (lazily
  mounted like every other wrapper; main.py shrinks well under the
  512-line cap)
- bare 'scitex completion' now shows group help instead of silently
  auto-installing
- 'completion install' gains --dry-run: prints the target rc file and
  the completion script without touching the filesystem (subsumes the
  old bash/zsh/fish script-dump leaves)
- 'completion bash|zsh|fish' become hidden warn-phase deprecated
  leaves: still print the script, warn on stderr, removed in v3.0
- root help examples updated ('scitex completion --show' never existed)
